packages/test-network: move types.ts
